+++ blaise.cli/Blaise.Cli.Core/Blaise.Cli.Core.csproj
@@ -150,12 +150,12 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Command\CommandParser.cs" />
+    <Compile Include="Services\CommandService.cs" />
     <Compile Include="Extensions\ArgumentValidationExtensions.cs" />
     <Compile Include="Interfaces\IBlaiseFileService.cs" />
-    <Compile Include="Interfaces\ICommandParser.cs" />
-    <Compile Include="Command\Models\DataDeliveryOptions.cs" />
-    <Compile Include="Command\Models\DataInterfaceOptions.cs" />
+    <Compile Include="Interfaces\ICommandService.cs" />
+    <Compile Include="Models\DataDeliveryOptions.cs" />
+    <Compile Include="Models\DataInterfaceOptions.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Services\BlaiseFileService.cs" />
   </ItemGroup>
