+++ blaise.cli/Blaise.Cli.Core/Interfaces/ICommandService.cs
@@ -1,6 +1,6 @@
 ﻿namespace Blaise.Cli.Core.Interfaces
 {
-    public interface ICommandParser
+    public interface ICommandService
     {
         int ParseArguments(string[] args);
     }
