+++ blaise.cli/Blaise.Cli.Tests.Unit/Blaise.Cli.Tests.Unit.csproj
@@ -73,10 +73,9 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Command\ParseDataDeliveryCommandsTests.cs" />
+    <Compile Include="Services\CommandServiceTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Services\BlaiseFileServiceTests.cs" />
-    <Compile Include="Command\ParseDataInterfaceCommandsTests.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="app.config" />
